Added doc/architecture.md
Mostly presenting this as a PR for comment. Trying to get some ideas
down in a semi-formal way in order to expose places we need clearer
address. Notably, ideas about how the server resolves the GDM and
unifying how we refer to artifacts need a little discussion. Could
happen in comments or back and forth commits here