Building up to a full integration test
Added exploratory/docker_labels - little tool to get the labels on an
image from its registry. Currently does not assume hub.docker.io or the
'latest' tag, so you'll need to spell all that out.

Determined the requirements for running a docker-composed registry -
needs to include a well-crafted self-signed cert, which will need to be
regenerated based on the IP address that docker-machine assigns. Will
next build a "rebuild" script for the cert and registry image - longer
term it'll be worthwhile checking that the cert matches the IP and only
rebuild when needed.