ADS test is skippable
If you run `go test ./... -short`, the ADS test will be skipped.

Also added: you can override the name of the docker-machine for testing
by using an environment variable.

For the moment test_with_docker has been moved into this repo. It seems
pretty robust and useful in other projects, but it remains easier to use
in one place.