Significant fixes to hy
There is a tedious but complicated issue with how hy Unmarshalled
structs if there was a hy tag on a field that wasn't the root object.
Among other things, this means that unmarshalling a struct that includes
another struct would behave differently in regards to the included
struct than unmarshalling the included struct directly.

n.b.: This is *still applies* to marshalling structs, but wasn't urgent
for Sous yet.
DCOPS-7242