Not the bugfix we deserve, but the bugfix we need
Holy bughunt, batman!

So: the "harvest" functionality was failing because docker's
representation because the repository name didn't have a digest or a tag
(and well it shouldn't - if it did it would be an image name...)

The Global inheritence wasn't completely impletmented.

Both now fixed with better tests and increased logging.

Would like to merge this and have Wil try it out