Adds "repo guessing" to the docker name cache
When the name cache is checked for the image name for a source ID
and there's no image name known,
it gets all tags from an appropriate repo.
If it doesn't know of an appropriate repo,
it now guesses, based on the algorithm that
the docker builder uses to generate a repo.

TECHOPS-1314