Sous resolver determines image name_resolver
This kind of exploded in scope. The idea was to push the guard for the
availability of images from the top level (and blocking the whole
rectification if it fails) to a per-deployment scope. We need to resolve
the image name to do that, and since that can involve a docker-registry
roundtrip, I wanted to store the retrieved value.

The result is that the deployer now works on "Deployables" as opposed to
"Deployments" - a Deployable is a pair of a Deployment and a
BuildArtifact. The top level "resolver" farms out the resolution of the
BuildArtifacts.

One very nice result is that the Singularity Deployer is now decoupled
from the Docker NameCache.

TECHOPS-1316