Fixes multiple cluster testing
The Skip in TestGetRunningDeploymentSet_all confused me, and in tracking
down the cause, I was able to fix it: the test container registration
didn't properly name the requests to be matched by the
deployment_builder.

It's still probably a better idea to be tagging Sing. Deploys with
metadata for this kind of thing, but it's fixed for the moment