Future-proofing the /servers API
In the future, we'll likely want to give cluster names to the returned
servers. We can change the configs pretty easily, but the deployed
client base will be harder to change. Wrapping the URL string in a
struct (-> a JSON object) lets us add e.g. a Name field when that's
needed.