Edge cases for status checks
Pollers aren't created for clusters that don't have a deployment
assigned to them.

Errors returned in the ResolveStatus are checked to see if there's a
chance they'll clear; if not, we treat the failure as permanent.

ResolveStatus got some extra values, including a MAX (so that the tests
can confirm there's a string value for everything without knowing what
the "top" status number is) and TERMINALS (which allows for multiple
end-state cases.)

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
# On branch bugfix_status_edgecase
# Changes to be committed:
#	modified:   lib/auto_resolver_test.go
#	modified:   lib/deployable_chans_test.go
#	modified:   lib/http_state_manager_test.go
#	modified:   lib/resolve_errors.go
#	new file:   lib/resolve_errors_test.go
#	modified:   lib/status_poller.go
#	modified:   lib/status_poller_test.go
#	modified:   lib/volumes_test.go
#	modified:   test/http_state_manager_test.go
#