Several enhancements and fixes.
(sorry this is scattered - juggling with office work & meetings)

sous update, if run in a properly tagged workspace, no longer requires a
-tag or -repo flags.

A significant issue with status polling is fixed: the status endpoint
handler reports the GDM as known to the AutoResolver, rather than the
GDM loaded at boot time. This bug was causing spurious ResolveNotVersion
results, because servers were reporting that they did not know of the
version we were waiting on - in fact they did, but they were reporting a
snapshot of the GDM from when they booted.

TECHOPS-1703