Hugely frustrating path problem
OSX uses a symlinked directory for its temporary directory (i.e. where a
reasonable OS would use /tmp, OSX does this /var/folders/... thing, but
/var is a symlink to /private/var)

We determine the root directory of a git repo by asking git, which
(conservatively) returns the physical directory (i.e. /private/var/...).
Then we try to get the relative path of the working directory based on
os.Getwd or the like (i.e. /var/...) The result is a relative directory
with a ladder of ../../... in it.

This commit updates our logic to comb through the prefix of the working
directory for symlinks that lead back into the root directory, and then
form a relativate path from there.

The goal is that symlinks above the git root will get resolved, but
symlinks below will be treated appropriately.