Making test for OS X tempdir conundrum OS indepentant
n.b.: This is a real issue. A dev might well symlink their dev
environment for any number of reasons, and we still need to recognize
that relative paths are within the git workspace. Conversely, a team
might make links within a repo for whatever reason, and we shouldn't
unpick those when naming Sous packages