Updates sous_qa_setup
Upshot is that *docker* treats 127.0.0.1 as a trustworthy registry, and
communicates with it over HTTP, so no certificate stuff is required.
Sous has a wonky workaround (BecomeFoolishlyTrusting) in its registry
client, which covers that communication channel.

On docker-machine, the registry is at a Virtualbox 192.168.0.0/16 address, so
(again) *docker* requires TLS, and needs a cert installed to verify the
registry. (BecomeFoolishlyTrusting has covered this case for
Sous-registry comms.)

We'll need to be able to supply a CA cert to our registry client, which
means that the testing.{crt,key} need to be generated and injected into
the registry regardless of the agent type.

Change is like this: we always (or will always) need certificates
*generated* for the test registry, but local_daemon setups don't need
them installed. So, sous_qa_setup generates them (unless the existing
one already covers the agentIP) and then has the agent "Install" them.

Local daemon agents treat the Install step as a no-op with an
instructive message about the situation.