Offset and Flavor now treat "" as a matchable value
That is the zero string is a legitimate value for Offset and Flavor, and
needs to be treated as a matching target - if the user (implicitly?)
tries to match Offset = "", that must not match "" and "something". The
UI is updated to accept '*' to mean "match anything" in these cases.

TECHOPS-1823