Corrects search for failed deploys
For requests with neither active nor pending deploys, we need to do one
extra HTTP request to get their recent deploy history.

On the one hand, this GET is unaviodable: we don't know the current
deploy ID without it. On the other, the fraction of all Requests that
will require this request should generally be low - since it's only
Requests that are in a failed state.