Requests limited to only legal characters
As restricted by Kafka, Singularity and Mesos. Docker is implied by
Singularity, afaict, but probably worth checking that as well.

Note that there's a disambiguation suffix added (ala deploy ids), but I
think that'll be minimal impact because request IDs get cut off in the
web UI as it is.