Removed Args field from DeployConfig
Args was part of lib/deploy_config.go when it was created in July, '16.
No GDM manifest makes use of this field.
It doesn't get translated into Singularity API calls at all.
It's effectively dead code at the moment.

If we need Args, they should be well defined as to what they do
i.e.
I think they're supposed to be passed as dockerParameters,
effectively flags to `docker run`
but I'm not 100% sure,
or 100% sure that's what we would want,
or if we would want that at all.

Increasingly,
I think that service configuration
should be passed as environment variables.
There's an exception to be made for
(iirc)
certain apps that need elevated permissions,
but that's it's own ball of wax,
and making that available
to all and sundry
seems like a bad idea.