Canary attributes
Old clients who aren't cool enough to retain fields will get 400 errors
now when they try to do PUTs.

The server adds an <etag>: "canary" pair to JSON responses and expects a
matching field to be in updates. While it would be possible to construct
the field and add it, it's simpler to just handle JSON properly, since
the field will be retained by the putbackJSON mechanism.