Adds a "SkipReadyCheck" field to Startup
Services that don't have a healthcheck endpoint can set SkipReadyCheck
in order to communicate that they shouldn't be probed with HTTP or TCP
on boot.

Specifically setting this up such that those services will need to set
it explicitly, in order to exert gentle pressure towards setting up a
startup check.

TECHOPS-2491