Logging infrastructure was messy
Frankly, it still is, but now it's less messy.

Specifically:
Graphite doesn't attempt to connect if it's not configured.

We log configuration re Kafka and Graphite, so that we can be sure at
boot time that we're actually making the connections as configured.

The graphite error message didn't implement the LogMessage interface
correctly and was being dropped. There's a regression test.

Build logs a specialized message to make use of the ConsoleMessage
interface when it selects a build strategy.