Fixed pipeline bug
The resolver err copying was pulling off of the *namer* Errs channel.
But so was the logging pipeline segment,
so there was a race for the DiffResolutions,
which was why integration tests only sometimes failed.
In the cases where the tests succeeded
(i.e. errs were reported at the end of the Resolve)
those errors wouldn't have been logged.