Defensive programming for panics
There's a test now for the diff messages, such that they omit fields
rather than drill down into nil fields (and thereby panic.)

Futhermore, there's a rescue in Deliver - if there's a panic in the
course of delivering a message, we log a (well tested and simple)
silentMessageErr - which will report the type of the "silent message":
loggingPanicFakeMessage, which is a signal that a log message panicked.

We'll get a log entry to mark the panic, but Sous won't crash because
formatting code related to producing a log message panicked.

SOUS-728