main: Fix -d and -v flags (guess verbosity early)
LogSet constructor needs to depend on verbosity to be configured
correctly. Before this change we could not know the verbosity until
flag.Parse which can't be called early enough to set up logging in main.

Therefore, we now sniff os.Args for -d or -v flags, and initialise
logging using that flag to define verbosity.

This change alone breaks reading Logging.Basic.Level from the config
file, so the next commit will address that.